# api/openapi.yaml
openapi: 3.1.0
info:
  title: (S)AGE API
  version: 3.5.0
  description: >-
    (Sovereign) Agent Governed Experience — REST API for memory submission,
    validation, and query.
  contact:
    name: Dhillon Andrew Kannabhiran
    email: dhillon@levelupctf.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security — Ed25519 signature authentication via headers
# ---------------------------------------------------------------------------
security:
  - ed25519Auth: []
components:
...
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >-
        Ed25519 authentication is conveyed through three headers sent together:
        **X-Agent-ID** (hex-encoded public key), **X-Signature** (Ed25519
        signature of the canonical request payload), and **X-Timestamp** (unix
        epoch seconds). All three must be present on authenticated endpoints.

  # -----------------------------------------------------------------------
  # Parameters
  # -----------------------------------------------------------------------
  parameters:
    memoryId:
      name: memory_id
      in: path
      required: true
      description: Unique memory object identifier (UUID v7).
      schema:
        type: string
        format: uuid

  # -----------------------------------------------------------------------
  # Schemas
  # -----------------------------------------------------------------------
  schemas:
...
    MemoryType:
      type: string
      enum:
        - fact
        - observation
        - inference
      description: Classification of the memory object.
    MemoryStatus:
      type: string
      enum:
        - proposed
        - validated
        - committed
        - challenged
        - deprecated
      description: Lifecycle state of a memory object.
    VoteDecisionEnum:
      type: string
      enum:
        - accept
        - reject
        - abstain

    # -- Core domain objects -----------------------------------------------
    KnowledgeTriple:
      type: object
      required:
        - subject
        - predicate
        - object
      properties:
        subject:
          type: string
          description: Subject entity of the triple.
        predicate:
          type: string
          description: Relationship between subject and object.
        object:
          type: string
          description: Object entity of the triple.
    MemoryRecord:
      type: object
      required:
        - memory_id
        - submitting_agent
        - content
        - content_hash
        - memory_type
        - domain_tag
        - confidence_score
        - status
        - created_at
      properties:
        memory_id:
          type: string
          format: uuid
          description: Unique identifier (UUID v7).
        submitting_agent:
          type: string
          description: Hex-encoded Ed25519 public key of the proposing agent.
        content:
          type: string
          description: Natural-language memory content.
        content_hash:
          type: string
          description: SHA-256 hash of content (hex-encoded).
        embedding:
          type: array
          items:
            type: number
            format: float
          description: 1536-dimensional embedding vector.
        embedding_hash:
          type: string
          description: SHA-256 hash of the embedding vector (hex-encoded).
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        domain_tag:
          type: string
          description: Domain or topic tag for the memory.
        confidence_score:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Confidence score assigned by the proposing agent.
        status:
          $ref: "#/components/schemas/MemoryStatus"
        parent_hash:
          type: string
          description: Hash of the parent memory (for lineage), if any.
        knowledge_triples:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeTriple"
          description: Extracted knowledge graph triples.
        votes:
          type: array
          items:
            type: object
            properties:
              voter_id:
                type: string
              decision:
                $ref: "#/components/schemas/VoteDecisionEnum"
              rationale:
                type: string
              voted_at:
                type: string
                format: date-time
          description: Votes cast on this memory.
        corroborations:
          type: array
          items:
            type: object
            properties:
              agent_id:
                type: string
              evidence:
                type: string
              created_at:
                type: string
                format: date-time
          description: Corroborations supporting this memory.
        created_at:
          type: string
          format: date-time
          description: Timestamp when the memory was proposed.
        committed_at:
          type: string
          format: date-time
          description: Timestamp when the memory reached committed status.

    # -- Request bodies ----------------------------------------------------
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.

  # -----------------------------------------------------------------------
  # Reusable responses
  # -----------------------------------------------------------------------
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/memory/{memory_id}:
    get:
      operationId: getMemory
      summary: Retrieve a single memory by ID.
      tags:
        - Memory
      parameters:
        - $ref: "#/components/parameters/memoryId"
      responses:
        "200":
          description: Full memory record with votes and corroborations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...